improving file structure
