+++ packages/frontend/src/components/index.ts
@@ -11,5 +11,4 @@ export * from "./name_input";
 export * from "./panel";
 export * from "./resizable";
 export * from "./rich_text_editor";
-export * from "./json_export";
 export * from "./json_import";
