+++ packages/frontend/src/page/menubar.tsx
@@ -5,17 +5,13 @@ import { useAuth, useFirebaseApp } from "solid-firebase";
 import { type JSX, Show, createSignal, useContext } from "solid-js";
 import invariant from "tiny-invariant";
 
-import type { AnalysisDocument } from "../analysis";
-import { type DiagramDocument } from "../diagram";
-import { type ModelDocument } from "../model";
-
 import { useApi } from "../api";
-import { Dialog, IconButton } from "../components";
 import { createModel } from "../model/document";
 import { TheoryLibraryContext } from "../stdlib";
-import { createDiagram } from "../diagram";
+import { Dialog, IconButton } from "../components";
 
 import { Login } from "../user";
+import { Open } from "./import";
 
 import FilePlus from "lucide-solid/icons/file-plus";
 import Info from "lucide-solid/icons/info";
@@ -24,10 +20,8 @@ import LogOutIcon from "lucide-solid/icons/log-out";
 import MenuIcon from "lucide-solid/icons/menu";
 import SettingsIcon from "lucide-solid/icons/settings";
 
-import { JsonImport } from "../components/json_import";
 import "./menubar.css";
 
-type Document = ModelDocument | DiagramDocument | AnalysisDocument;
 /** Menu triggered from a hamburger button. */
 export function HamburgerMenu(props: {
     children: JSX.Element;
@@ -177,62 +171,5 @@ function OpenMenuItem(props: {
             <LogInIcon />
             <MenuItemLabel>{"Open notebook"}</MenuItemLabel>
         </MenuItem>
-    );
-}
-
-function Open(props: { onComplete?: () => void }) {
-    const api = useApi();
-    const navigate = useNavigate();
-    const handleImport = async (data: Document) => {
-        console.log("Imported data:", data);
-
-        switch (data.type) {
-            case "model": {
-                const newRef = await createModel(api, {
-                    ...data,
-                    name: `${data.name}`,
-                });
-                navigate(`/model/${newRef}`);
-                break;
-            }
-            // XX: Probably won't work yet
-            case "diagram": {
-                const newRef = await createDiagram(api, {
-                    ...data,
-                    name: `${data.name}`,
-                });
-                navigate(`/diagram/${newRef}`);
-                break;
-            }
-
-            case "analysis": {
-                throw new Error("Analyses don't currently support initialization.");
-            }
-
-            default:
-                throw new Error("Unknown document type");
-        }
-
-        props.onComplete?.();
-    };
-
-    // Placeholder, not doing more than typechecking does for now but
-    // will eventually validate against json schema
-    const validateJson = (data: Document) => {
-        // Return true if valid
-        if (data.name && data.notebook && data.type) {
-            return true;
-        }
-        // Return error message if invalid
-        return 'JSON must include "name", "notebook", and "type" fields';
-    };
-
-    return (
-        <div>
-            <JsonImport
-                onImport={handleImport}
-                validate={validateJson} // Optional
-            />
-        </div>
     );
 }
